CU-2atcgy0 - Creat migration and model for attachments - Get fill uploader and make into livewire component - Work on attachment functionality in the backend - Add attachment component to new post box

+++ Modules/Social/Database/Migrations/2022_02_17_221205_create_posts_table.php
@@ -18,7 +18,7 @@ class CreatePostsTable extends Migration
             $table->foreignIdFor(\App\Models\User::class, 'user_id');
             $table->unsignedBigInteger('team_id')->nullable();
             $table->text('body');
-            $table->integer('postable_id')->nullable();
+            $table->unsignedBigInteger('postable_id')->nullable();
             $table->string('postable_type')->nullable();
             $table->timestamps();
         });
